Don't always copy network by default
If no network data is supplied by Ignition and the host is PXE booted, a
network config is created that prevents any interfaces other than the
PXE interface from coming up. If this is then copied to the installed
system with --copy-network then only the PXE interface will get an IP
address when the host comes up:
https://bugzilla.redhat.com/show_bug.cgi?id=2032717

Since not customising the network is the default, only pass
--copy-network by default if the environment variable
IPA_COREOS_COPY_NETWORK is set to "true". This should be set in the
ignition if network customisation is present (making it explicit avoids
the custom install step needing to read the Ignition). The default can
still be overridden by the 'coreos_copy_network' field in the metadata.

https://bugzilla.redhat.com/show_bug.cgi?id=2038303